web_app/requirements.txt
httpx==0.21.0
fluent-logger==0.10.0
fastapi==0.95.*
uvicorn[standard]==0.21.*